backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pyswisseph==2.10.3.2
python-dateutil==2.8.2
pytz==2023.3
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2